Split TX lifecycle module
